+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Include `pickleId` in `AssembledTestCase` ([#26](https://github.com/cucumber/javascript-core/pull/26))
 
 ## [0.5.1] - 2025-09-17